requirements.txt
PyYAML~=6.0.1
transformers~=4.36.0
torch~=2.1.1
peft~=0.6.2
setuptools~=65.5.1
hydra-core==1.2.0
omegaconf~=2.3.0
fastapi~=0.109.1
pydantic~=2.5.1
uvicorn~=0.24.0.post1
openai==1.7.2
sentence_transformers
requests~=2.31.0
aleph-alpha-client==7.0.0
pytest~=7.4.4